rate limit

+++ config.go
@@ -41,7 +41,7 @@ type Config struct {
 	IPBlocklist iplist.Ranger
 	DisableIPv6 bool `long:"disable-ipv6"`
 	// how many kB can be send every second
-	SendPieceRate uint32 `long:"max-kbyte-can-send-every-second"`
+	SendPieceRate int64 `long:"max-kbyte-can-send-every-second"`
 	// Perform logging and any other behaviour that will help debug.
 	Debug bool `help:"enable debug logging"`
 }
